fix compiler warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@22929 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/nas/nashandler.cpp
@@ -134,7 +134,7 @@ void NASHandler::startElement(const XMLCh* const    uri,
         if( m_pszGeometry == NULL )
             m_nGeometryDepth = poState->m_nPathLength;
         
-        if( m_nGeomLen + nLNLen + 4 + osAttributes.size() > m_nGeomAlloc )
+        if( m_nGeomLen + nLNLen + 4 + (int)osAttributes.size() > m_nGeomAlloc )
         {
             m_nGeomAlloc = (int) (m_nGeomAlloc * 1.3 + nLNLen + osAttributes.size() + 1000);
             m_pszGeometry = (char *)
